Add legal controls to lock screen
